Initial bandwidth metric events
